feat: converter to sha256
Because C# doesn't have easy way to convert strings to sha256 so we will need a function for it. This function will use as an extension method. Hashing in C# has some parts: 1.create a instance of hashing algorithm; 2. Encode with a algorithm; 3. Do the main hash; 4. Convert the result to string