feat: store time
We need to stores the amount of time our app wants to happen again and again. Besides that we can see we are using a unusual Using part. This way of Using is because the Timer class is a same class in two library. According to Tim's video both classes which are System.Timers and System.Threading work basically the same but the multi threading is also here in TImers so we will use the Timers class. But because the vscode keep shows us error i decided to use that class like this so we won't ger error anymore