feat [manage constructor]: make an event
We will make an event for time when the time elapsed so we append a event method to this prop. We can make the method name and use the ctrl + dot to make our event function automatically.
There is also a thrown object which Tim mentioned. This NotImplementedException method will cause exception if we call it some where so we can notice that this method didn't completed