chore [Retry]: additional files
Because the designing of this project was wrong I should start it from start but i chose to keep some part of code and recommit other parts. It's like nightmare but this is the way of learning. It's my first project in dotnet core and vscode so it's kind of normal :smile: