feat: hide user password
Same as Linux validation it's better to not should the user password. This method also has bug. When user type things and then start to backspace it, the asked string starts to delete from console. In next version of Spectre.Console the authors will solve this problem, they wrote it already