feat: handle unwanted input
This method will run then the user insert an invalid input e.g. user should insert a number but he insert a char, so here this method runs and notify the user about his wrong input. In our code this method would never run because we want string and also the input in normal format is string, but I will keep it because it would be useful in future