feat [validation]: correct password
When user insert a string we will convert it to sha256 and compares it with sha256 of saved password. Here as test I set the second string test but it should replaced with a method in configHandling.RootPart which returns the root password in hash format. Here also we return the success function which means validation completed successfully