feat: not same recovery passwords
When passwords goes wrong a message shows which notify user that passwords wasn't the same and press a key to repeat the process. It also waits for a key from user and at the end clears the console for menu. Because this process is inside of a while loop this process repeats until same passwords inserted