feat [Menu]: add menu prompt handler
This method will handle menu prompt of Spectre.Console. Beside of that it's public, it's static which means no needs for calling its constructor. It also returns a list of object which means we could put any type of items in that list. This method only wants a list of strings which are names or rows of menu