feat: create method for password prompt
This method beside of being public, it's static. The parameters are a name for password which could password or recovery password, IsNew boolean which shows that this method calls for first time of used after root password has been set. It has also a NeedsHint boolean which is used for recovery passwords which needs a hint