feat: use AnsiConsole prompt
This prompt will show a message to user and ask for a string. The user input style sets to green. It also hide user password from seeing, and also shows a message when user insert a different format from string which here is kind of useless but could be helpful