feat: log file path
When this property get use because of it's static, it will set for other part of codes too. beside it's read only, it runs the pwd command to found where it is in user device, then does some normaliztion on it which are repalce some names and etc. This normaliztion does with regex lib