feat: create date method
This method will returns date of log files when we call it. It can have input string which help us decide what we should do, return date of first log or return date of given string. Here we using regex and also the method itself is static for easier use