fix: config file path
Because both service part and manage part should be able to work with config file and in each of them the config file path goes different. I tried multiple ways but this is what handles it so well. First it gets the pwd command result which let us know where the files are. Then it replaces some of words of directories with regex and sets it as path in configFilePath property