feat: finished time
After we log a time we check it user has time to use or not. I use a customized method which name is Allowed. This method will let us know that the user has time to spend or not. If time has ended it prints time has ended and then stops the service