feat: set service part config string
The service part config string will contains 9 parts which are separated from each other with semicolon. The config string also has hash of 0 as shown for parts status which are off at the first time. The string has some default values which aren't used and used for making the file less understandable