feat: add method
This method do things like normal content method I created already but it has a different. This method will return part by part of config string we give to it not read config file and split it. Same as other method it first remove '{' and '}' and then separate parts by ';'. At the end we have array of string which each index represents each part of config string