feat: add header
By calling the header method in all parts of our code we will have cool header in all parts of our code. It also should be inside loops because of repeatation of that part for errors. Using multiple time of header method also doesn't make problem because the header method clears the console before add the line. cool :)