feat: change exit logic
The way of exiting wasn't same as other ways in manage app so I changed it. Now on user should select Yes or No in menu. By calling the menuPrompt method we pass the options in a array of string and a message in custom format. It also should be use in a switch to handles it. The value we should handle is a int which is the row number of menu